% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{check_status}
\alias{check_status}
\title{check_status()}
\usage{
check_status()
}
\description{
Function to query the CAT Public API and check the status of CATSERVER.
Handy helper function that, if you run before using the optimizer,
will ensure that the API is warmed up and ready to go for you.
}
\author{
Tim Fraser
}
