% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{check_values_diversification}
\alias{check_values_diversification}
\title{Helper Function: Check Values from Diversification Metric}
\usage{
check_values_diversification(n = 1000, dmax = 5, n_policies = 3)
}
\description{
Takes many random samples of policy scales and computes diversification values for that compilation of values.
Allows you to see the range of diversification values.
}
