% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{connect}
\alias{connect}
\title{functions_evaluation.R}
\usage{
connect(type = "policies")
}
\arguments{
\item{type}{(character) database name. Choose from `"policies"`, `"cost_effectiveness"`, `"granddata"`, `"geo"`}
}
\description{
This script develops the policy optimizer's functions for evaluating metrics.
This produces the starter effects data used.

Internal function for connecting to databases.
}
