% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{get_avg}
\alias{get_avg}
\title{get_avg}
\usage{
get_avg(data, var = "ce", range = 5)
}
\description{
Get n-year averages for cost effectiveness statistics or emissions per dollar statistics.
}
