% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{get_catdata}
\alias{get_catdata}
\title{`get_catdata()`}
\usage{
get_catdata(input, test = FALSE)
}
\arguments{
\item{input}{list of `geoid`, `pollutant`, and `by` aggregation type. Will be necessary for querying catserver eventually.}

\item{test}{(logical) is this a test? If so, just return `catdata_test`. Otherwise, actually query the catserver, which requires a `.Renviron` file}
}
\description{
Function to query catserver. Currently just relies on catdata_test
}
\note{
Depends on catviz package
}
\author{
Tim Fraser
}
