% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{get_constants}
\alias{get_constants}
\title{Get Constants used in Optimization}
\usage{
get_constants(
  effects,
  training,
  n_scales = 100,
  .year = 2005,
  .percent_change = 0.5
)
}
\description{
Get Constants used in Optimization
}
