% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{get_cost_effectiveness}
\alias{get_cost_effectiveness}
\title{get_cost_effectiveness}
\usage{
get_cost_effectiveness(
  data = read_csv("data_raw/ce_averages.csv"),
  .pollutant = 98,
  .category_id = 1,
  .start_year = 2020,
  .end_year = 2025
)
}
\description{
For any pollutant,
For any number of policies,
For any year range,
Get cost effectiveness with the cost model
}
