% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{get_cost_model}
\alias{get_cost_model}
\title{get_cost_model}
\usage{
get_cost_model(
  data = read_csv("data_raw/ce_averages.csv"),
  .pollutant = 98,
  .category_id = 1
)
}
\description{
get_cost_model
}
