% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{get_cost_model_simple}
\alias{get_cost_model_simple}
\title{get_cost_model_simple}
\usage{
get_cost_model_simple(
  data = read_csv("data_raw/ce_averages.csv"),
  .pollutant = 98,
  .category_id = 1
)
}
\description{
A cost model where we take the five year average
}
