% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{get_diversification}
\alias{get_diversification}
\title{Diversification Function for 1 year}
\usage{
get_diversification(absolute, n, width, prior_diversification = 0)
}
\description{
Measure diversification in 1 year, by taking the sum of absolute deviations, normalized from 0 to 1.
}
\author{
Tim Fraser
}
