% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{get_effect_ef}
\alias{get_effect_ef}
\title{Get Effect on Emissions Factor (tons / mile) in given year}
\usage{
get_effect_ef(effect_ef, units, d, n)
}
\description{
Simple function for getting annual effect on EF, just for record keeping.
}
\author{
Tim Fraser
}
