% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{get_effect_emission}
\alias{get_effect_emission}
\alias{get_effect_emissions}
\title{Get Effect on Emissions in given year}
\usage{
get_effect_emissions(effect_emissions, units, d, n)
}
\description{
Simple function for getting annual effect on emissions, just for record keeping. 
See `get_reduction()` for the sub-function used in the objective function.
}
\author{
Tim Fraser
}
