% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{get_effect_vmt}
\alias{get_effect_vmt}
\title{Get Effect on VMT in given year}
\usage{
get_effect_vmt(effect_vmt, units, d, n)
}
\description{
Simple function for getting annual effect on VMT, just for record keeping.
}
\author{
Tim Fraser
}
