% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{get_effects}
\alias{get_effects}
\title{get_effects}
\usage{
get_effects(
  .geoid = "36109",
  .pollutants = c(2, 3, 87, 98, 100),
  var = "epd",
  path = NULL,
  prob = 0.5,
  min_cost = 0,
  range = 5,
  .policy = c(1, 2),
  .start_year = 2020,
  .end_year = 2025,
  .units = 30000,
  fundamental = FALSE,
  n = 1000
)
}
\description{
get_effects
}
