% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{get_frame}
\alias{get_frame}
\title{`get_frame()`}
\usage{
get_frame(data)
}
\description{
Function to get `catdata` interpolated to cover every year in the year-range supplied
}
\author{
Tim Fraser
}
