% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_reporting.R
\name{get_pollutant}
\alias{get_pollutant}
\title{functions_reporting.R}
\usage{
get_pollutant(pollutant = 98)
}
\arguments{
\item{pollutant}{(integer) EPA pollutant code. `98` is CO2 equivalent emissions.}
}
\description{
This script develops the policy optimizer's functions for reporting outcomes.
This includes reporting statistics, visualizations, and summary tables.
}
