% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{get_record}
\alias{get_record}
\title{Get Record data.frame used in Optimization}
\usage{
get_record(
  geoid = "36109",
  start_year = 2025,
  end_year = 2050,
  units = 30000,
  policies = 1:10,
  pollutants = c(2, 3, 87, 98, 100),
  last_cost = 0,
  last_reduction98 = 0,
  last_reduction2 = 0,
  last_reduction3 = 0,
  last_reduction87 = 0,
  last_reduction100 = 0
)
}
\description{
Construct a data.frame `record` that records values over time. 
Output has 1 row per year in the scenario, plus placeholders for metrics for each yer.
}
\author{
Tim Fraser
}
