% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{get_reduction}
\alias{get_reduction}
\title{Emissions Reduction Function}
\usage{
get_reduction(effect_emissions, units, d, n, prior_reduction = 0)
}
\arguments{
\item{effect_emissions}{(numeric) value of policy effects on emissions, of length `n`}

\item{units}{(integer, length 1) constant value for up-scaling scenario implementation level. 30,000 seems to do well for counties.}

\item{d}{decision object from ompr}

\item{n}{(integer, length 1) number of policies under consideration.}

\item{prior_reduction}{(double, length 1) cumulative emissions reduction from prior year.}
}
\description{
Emissions Reduction Function
}
\author{
Tim Fraser
}
