% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{get_z}
\alias{get_z}
\title{Objective Function}
\usage{
get_z(cost = 1, reduction = 1, wc = 0, wr = 0)
}
\arguments{
\item{cost}{(numeric) cost metric, from 0 (best) to 1 (worst)}

\item{reduction}{(numeric) reduction metric, from 0 (best) to 1 (worst)}

\item{wc}{(percent) weight of cost metric, from 0 (least weight) to 1 (greatest weight)}

\item{wr}{(percent) weight of reduction metric, from 0 (least weight) to 1 (greatest weight)}
}
\description{
Combine scaled cost and scaled emissions reduction metric using weights `wc` and `wr`.
}
\author{
Tim Fraser
}
