% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interal_functions_FasFran.R
\name{getDstat}
\alias{getDstat}
\title{Get KS Stat}
\usage{
getDstat(originSamples, S1, S2, cores = 1)
}
\arguments{
\item{originSamples}{a \verb{[n by 2]} \code{data.frame} of x and y coordinates that defines the origins data points.}

\item{S1}{a \verb{[n by 2]} \code{data.frame} of x and y coordinates of sample 1}

\item{S2}{a \verb{[n by 2]} \code{data.frame} of x and y coordinates of sample 2}

\item{cores}{a \code{numeric} defining the number of cores to use of processing}
}
\value{
a \code{numeric} defining the D stat with the largest difference between the quad frequencies, after checking each point as the origin
}
\description{
Loop through each row as(i.e. data point)
and defines D stat as the largest difference between the quadfrequencies, looping though each point in the sample as the origin
}
\details{
Code adapted from Press, W. H., Teukolsky, S. A., Vetterling, W. T.,, Flannery, B. P. (2007). Numerical Recipes 3rd Edition: The Art of Scientific Computing. Cambridge University Press. ISBN: 0521880688
}
