% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interal_functions_FasFran.R
\name{quadCount}
\alias{quadCount}
\title{Quad Count}
\usage{
quadCount(x_origin, y_origin, x, y)
}
\arguments{
\item{x_origin}{a \code{numeric} defining the x coordinate of the origin defining the 4 quadrants}

\item{y_origin}{a \code{numeric} defining the y coordinate of the origin defining the 4 quadrants}

\item{x}{a \code{vector} of \code{numeric} x coordinates}

\item{y}{a \code{vector} of \code{numeric} y coordinates}
}
\value{
a \code{vector} of frequencies of the number of points in each of the four quadrants defined by the origin point
}
\description{
Counts the frequency of points in the four quadrants - starting from the upper right going counter clockwise. Quadrants defined by the origin points x and y.
}
\details{
Code adapted from Press, W. H., Teukolsky, S. A., Vetterling, W. T.,, Flannery, B. P. (2007). Numerical Recipes 3rd Edition: The Art of Scientific Computing. Cambridge University Press. ISBN: 0521880688
}
