% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_reporting.R
\name{report_stats}
\alias{report_stats}
\title{report_stats}
\usage{
report_stats(record, pollutant = 98)
}
\description{
Calculates summary statistics for a `record` data.frame compiling the results for n counties.
Outputs a table of statistics and standard errors where applicable.
}
