% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tradespace.R
\name{sample_chull}
\alias{sample_chull}
\title{sample_chull}
\usage{
sample_chull(
  bundle,
  n = 10,
  fuzz = 5,
  n_scales = 100,
  record,
  effects,
  constants
)
}
\description{
sample_chull
}
