% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tradespace.R
\name{sample_outward}
\alias{sample_outward}
\title{functions_tradespace.R}
\usage{
sample_outward(
  seed,
  n_policies = 10,
  n_years = 26,
  n_scales = 100,
  fuzz = NULL
)
}
\description{
This script develops the policy optimizer's functions for sampling and visualizing the tradespace.
These do not directly affect the optimizer,
but are auxillary functions for comparison with the optimizer's findings.
}
