% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tradespace.R
\name{sample_scenario}
\alias{sample_scenario}
\alias{scenario_sampler}
\title{Get a Sample Scenario via random sampling}
\usage{
scenario_sampler(
  n = 1000,
  n_scales = 100,
  record,
  effects,
  constants,
  wc = 0.5,
  wr = 0.5,
  seed = NULL,
  fuzz = NULL
)
}
\arguments{
\item{n_scales}{(integer) max policy adoption scale}

\item{record}{(data.frame) a `record` data.frame object produced by `get_record()`. Gets filled in by this function.}

\item{effects}{(data.frame) an `effects` data.frame of policy effects over time corresponding to the years in your scenario. Produced by `get_effects()`.}

\item{constants}{(data.frame) a `constants` data.frame of constants produced by `get_constants()`. Used for rescaling metrics.}

\item{.id}{(integer) Default is NULL. If not null, the value to be printed at the conclusion of the sampling process. Used for progress reporting.}
}
\description{
This function generates a scenario of policy adoption scales at random 
and calculates the resulting `record` object.
Useful for enumerating an approximation of the full tradespace of your county's metrics.
}
\author{
Tim Fraser
}
