% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{scale_cost}
\alias{scale_cost}
\title{Scale Cost Metric}
\usage{
scale_cost(stat, lc, uc)
}
\description{
Function for rescaling cost metric using a lower cost and upper cost value.
Final metric will range from 0 (lowest cost) to 1 (highest cost)
}
\author{
Tim Fraser
}
