% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_optimization.R
\name{scale_reduction}
\alias{scale_reduction}
\title{Scale Emissions Reduction Metric}
\usage{
scale_reduction(stat, lr, ur)
}
\description{
Function for rescaling reduction metric using a lower reduction and upper reduction value.
Final metric will range from 0 (greatest reduction / lowest emissions) to 1 (lowest reduction / highest emissions)
}
\author{
Tim Fraser
}
