% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{sim_diff}
\alias{sim_diff}
\title{sim_diff}
\usage{
sim_diff(
  x0,
  x1,
  var = "proxy",
  model,
  adjust = -1,
  n = NULL,
  fundamental = FALSE
)
}
\description{
For a given estimation, get back differences with simulated standard errors
}
