% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_reporting.R
\name{summarize_by_policy}
\alias{summarize_by_policy}
\title{summarize_by_policy}
\usage{
summarize_by_policy(record, path = NULL)
}
\arguments{
\item{record}{(data.frame) optimization scenario record. Outputted by `policy_optimizer()`.}

\item{path}{(character) (Default = `NULL) (Optional) path to a cmaq `.csv` file. If not provided, use default. (Strongly advise you not change unless confident about it.)}
}
\description{
Summarize an optimization scenario `record` data.frame by policy effects.
}
