% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_reporting.R
\name{summarize_by_year}
\alias{summarize_by_year}
\title{summarize_by_year}
\usage{
summarize_by_year(record)
}
\arguments{
\item{record}{(data.frame) optimization scenario record. Outputted by `policy_optimizer()`.}
}
\description{
Summarize an optimization scenario `record` data.frame by policy outcomes over time.
}
