% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_reporting.R
\name{visualize_by_policy}
\alias{visualize_by_policy}
\title{visualize_by_policy}
\usage{
visualize_by_policy(data, var = "effect_cost", palette = "mako", adjust = 100)
}
\arguments{
\item{data}{data.frame generated by summarize_by_policy}

\item{var}{(character) a single variable, eg. `"effect_cost"` from your data.frame `data`}

\item{palette}{(character) viridis-style color palette name, eg. `"mako"`}

\item{adjust}{(numeric) adjustment for axis. Default is `100`.}
}
\description{
Visualize optimized scenario by policy, using data generated by `summarize_by_policy()`
}
