% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_reporting.R
\name{visualize_by_year}
\alias{visualize_by_year}
\title{visualize_by_year}
\usage{
visualize_by_year(data, var = "cost", palette = "mako", adjust = 100)
}
\arguments{
\item{data}{data.frame generated by summarize_by_year}

\item{var}{(character) a single variable, eg. `"cost"` from your data.frame `data`}

\item{palette}{(character) viridis-style color palette name, eg. `"mako"`}

\item{adjust}{(numeric) adjustment for axis. Default is `100`.}
}
\description{
Visualize optimized scenario by year, using data generated by `summarize_by_year()`
}
