% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_evaluation.R
\name{what_level}
\alias{what_level}
\title{Check What Level is my geoid}
\usage{
what_level(geoid)
}
\arguments{
\item{geoid}{(character) unique census geographic id for each county, state, county subdivision/municipality, nation, etc.}
}
\description{
Check What Level is my geoid
}
\author{
Tim Fraser
}
